requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
charset-normalizer==3.4.3
contourpy==1.3.3
cycler==0.12.1
fonttools==4.60.1
gigachat==0.1.42.post2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain-core==1.0.0
langchain-gigachat==0.3.12
langchain-text-splitters==1.0.0
langgraph==0.6.10
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.34
markupsafe==3.0.3
matplotlib==3.10.7
numpy==2.3.3
orjson==3.11.3
ormsgpack==1.11.0
packaging==25.0
pillow==11.3.0
pyan3==1.2.0
pydantic==2.12.0
pydantic-core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyyaml==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
threadpoolctl==3.6.0
types-requests==2.32.4.20250913
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
xxhash==3.6.0
zstandard==0.25.0C